feat(BE5011795.go): add HomeworkBE5011795 function to check if a string is a palindrome
This commit adds a new function called HomeworkBE5011795 to the mytesting package. The function takes a string as input and checks if it is a palindrome. If the string is a palindrome, the function returns true, otherwise it returns false. The function also returns an error if the length of the input string is less than 3 or greater than 15.