feat(BE4920086.go): add HomeworkBE4920086 function to sort an array of integers
This commit adds a new function called HomeworkBE4920086 that sorts an array of integers in ascending order. The function takes an array of integers as input and returns a sorted array of integers. The function also checks if the input array has less than 5 or greater than 30 numbers and returns an error if it does.